Removed ID from incentive apprenticeship dto

+++ src/SFA.DAS.EmployerIncentives.Database/Tables/IncentiveApplicationApprenticeship.sql
@@ -2,7 +2,7 @@
 (
 	[Id] UNIQUEIDENTIFIER NOT NULL PRIMARY KEY,
 	[IncentiveApplicationId] UNIQUEIDENTIFIER NOT NULL,
-	[ApprenticeshipId] INT NOT NULL,
+	[ApprenticeshipId] BIGINT NOT NULL,
 	[FirstName] NVARCHAR(100) NOT NULL,
 	[LastName] NVARCHAR(100) NOT NULL,
 	[DateOfBirth] DATETIME2 NOT NULL,
